requirements.txt
fastapi==0.115.12
sqlalchemy==2.0.40
alembic==1.15.2
pydantic-settings==2.9.1
uvicorn==0.34.2
email-validator==2.2.0
pwdlib[argon2]==0.2.1
pyjwt==2.10.1
python-multipart==0.0.20
psycopg[binary]==3.2.7
redis==6.1.0